# swagger.yml
swagger: '2.0'
info:
    description: This is the swagger file that goes with our server code
    version: '1.0.0'
    title: Swagger Rest Article
consumes:
    - application/json
produces:
    - application/json
basePath: /api
paths:
...
    /user/{id}:
        get:
            operationId: user.get_one
            tags:
                - user
            summary: Read one user from the user list
            description: Read one user from the user list
            parameters:
                - name: id
                  in: path
                  description: Id of the user to get from the list
                  type: integer
                  required: True
            responses:
                200:
                    description: Successfully read user from user list operation
                    schema:
                        properties:
                            id:
                                type: integer
                            username:
                                type: string
                            email:
                                type: string
                            phone:
                                type: string
                            country:
                                type: string
                            city:
                                type: string
                            postcode:
                                type: string
                            name:
                                type: string
                            address:
                                type: string

        put:
            operationId: user.update
            tags:
                - user
            summary: Update a user in the user list
            description: Update a user in the user list
            parameters:
                - name: id
                  in: path
                  description: Id of the user to update in the list
                  type: integer
                  required: True
                - name: user
                  in: body
                  schema:
                      type: object
                      properties:
                          username:
                              type: string
                              description: Username of user to update
                          email:
                              type: string
                              description: Email of user to update
                          password:
                              type: string
                              description: Password of user to update
                          phone:
                              type: string
                              description: Phone of user to update
                          country:
                              type: string
                              description: Country of user to update
                          city:
                              type: string
                              description: City of user to update
                          postcode:
                              type: string
                              description: Postcode of user to update
                          name:
                              type: string
                              description: Name of user to update
                          address:
                              type: string
                              description: Address of user to update
            responses:
                200:
                    description: Successfully updated user in user list

        delete:
            operationId: user.delete
            tags:
                - user
            summary: Delete a user from the user list
            description: Delete a user
            parameters:
                - name: id
                  in: path
                  type: integer
                  required: True
            responses:
                200:
                    description: Successfully deleted a user from user list